// front/src/shared/models/User.ts
export class User {
    firstname: string;
    lastname: string;
    adress: string;
    zipcode: string;
    city: string;
    gender: string;
    mail: string;
    password: string;
    login: string;
    country: string;
    phone: string;

    constructor () {
      // this.firstname = firstname;
      // this.lastname = lastname;
      // this.adress = adress;
      // this.zipcode = zipcode;
      // this.city = city;
      // this.gender = gender;
      // this.mail = mail;
      // this.password = password;
      // this.login = login;
      // this.country = country;
    }

  }